break linkedItemCard file into multiple files
